Whole bunch of refactoring, authentication fixes, added split view controller

+++ UFSVKFriendList/DataSource/DataManager.h
@@ -12,11 +12,11 @@
 
 + (DataManager *)sharedInstance;
 
-- (void)getFriendsForUserId:(long)userID offset:(NSInteger)offset count:(NSInteger)count
+- (void)getFriendsForUserId:(NSInteger)userID offset:(NSInteger)offset count:(NSInteger)count
                success:(void(^)(NSArray *friends))success
                failure:(void(^)(NSError *error, NSInteger statusCode)) failure;
 
-- (void)getGroupsForUserId:(long)userID offset:(NSInteger)offset count:(NSInteger)count
+- (void)getGroupsForUserId:(NSInteger)userID offset:(NSInteger)offset count:(NSInteger)count
                     success:(void(^)(NSArray *groups))success
                     failure:(void(^)(NSError *error, NSInteger statusCode)) failure;
